feat(download-upgrade): refactored code to download directly from gcp
